src/celery/requirements.txt
celery==5.2.7
openpyxl==3.1.0
python-dotenv==0.21.1